riotboard: enable usb_mon in defconfig
for unknown reasons, enabling CONFIG_USB_MON=m seems to help some usb wifi
adapters to work. This module doesn't seem to require to be loaded, just
building it as part of the config appears to be enough.

This is known to affect at least the netgear WNA1100 which is an
atheros 9271 chipset using the ath9k_htc driver.